update price course

+++ src/main/java/org/example/lmsbackend/dto/CourseDTO.java
@@ -1,12 +1,14 @@
 package org.example.lmsbackend.dto;
 
+import java.math.BigDecimal;
+
 public class CourseDTO {
     private String title;
     private String description;
     private int categoryId;
     private int instructorId;
     private String status;
-    private double price;
+    private BigDecimal price;
 
     // Getters
     public String getTitle() { return title; }
@@ -14,7 +16,7 @@ public class CourseDTO {
     public int getCategoryId() { return categoryId; }
     public int getInstructorId() { return instructorId; }
     public String getStatus() { return status; }
-    public double getPrice() { return price; }
+    public BigDecimal  getPrice() { return price; }
 
     // Setters
     public void setTitle(String title) { this.title = title; }
@@ -22,5 +24,5 @@ public class CourseDTO {
     public void setCategoryId(int categoryId) { this.categoryId = categoryId; }
     public void setInstructorId(int instructorId) { this.instructorId = instructorId; }
     public void setStatus(String status) { this.status = status; }
-    public void setPrice(double price) { this.price = price; }
+    public void setPrice(BigDecimal  price) { this.price = price; }
 }
